requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
flask-serialize==2.1.3
flask-cors==3.0.10
PyJWT==2.6.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
oauthlib==3.2.2
requests==2.29.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0